videoClassification/requirements.txt
torch
torchvision
tqdm
pandas
Pillow
matplotlib
numpy
